requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
dj-database-url==2.1.0
Django==4.2.11
django-autoslug==1.9.9
django-cleanup==8.1.0
django-storages==1.14.2
frozenlist==1.4.1
geoip2==4.8.0
gunicorn==22.0.0
idna==3.7
maxminddb==2.6.1
multidict==6.0.5
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
